pip_requirements.txt
absl-py==1.2.0
aiohttp==3.8.1
aiosignal==1.2.0
asttokens==2.0.5
astunparse==1.6.3
async-timeout==4.0.1
attrs==21.4.0
backcall==0.2.0
blinker==1.4
Bottleneck==1.3.5
brotlipy==0.7.0
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.1.0
click==8.0.4
cryptography==37.0.1
cycler==0.11.0
decorator==5.1.1
executing==0.9.1
flatbuffers==1.12
fonttools==4.25.0
frozenlist==1.2.0
gast==0.3.3
google-auth==2.9.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.42.0
h5py==2.10.0
idna==3.3
importlib-metadata==4.12.0
ipython==8.4.0
jedi==0.18.1
joblib==1.1.0
Keras-Preprocessing==1.1.2
keras-tuner==1.1.3
kiwisolver==1.4.2
kt-legacy==1.0.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
multidict==5.2.0
munkres==1.1.4
numexpr==2.7.3
numpy==1.19.2
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pip==22.1.2
ply==3.11
prompt-toolkit==3.0.30
protobuf==3.20.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.12.0
PyJWT==2.4.0
pyOpenSSL==22.0.0
pyparsing==3.0.4
PyQt5-sip==12.11.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.1.1
scipy==1.6.2
seaborn==0.11.2
setuptools==61.2.0
sip==6.6.2
six==1.15.0
stack-data==0.3.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.4.0
tensorflow-gpu==2.4.0
termcolor==1.1.0
threadpoolctl==2.2.0
toml==0.10.2
tornado==6.1
tqdm==4.64.0
traitlets==5.3.0
typing_extensions==4.1.1
urllib3==1.26.11
wcwidth==0.2.5
Werkzeug==2.2.1
wheel==0.37.1
wrapt==1.14.1
yarl==1.6.3
zipp==3.8.1
